Move relevant files/folders into data folder
- actions/
- reducers/
- store.ts
- types.ts
- utilities.ts
